requirements.txt
PyMuPDF
spacy
transformers
faiss-cpu
flask
flask-cors
pdfplumber
numpy
google-generativeai
gunicorn
python-dotenv